IGNITE-19025 Remove deployment code from API module
