+++ modules/client/src/main/java/org/apache/ignite/internal/client/TcpIgniteClient.java
@@ -27,7 +27,6 @@ import java.util.function.BiFunction;
 import org.apache.ignite.client.IgniteClient;
 import org.apache.ignite.client.IgniteClientConfiguration;
 import org.apache.ignite.compute.IgniteCompute;
-import org.apache.ignite.deployment.IgniteDeployment;
 import org.apache.ignite.internal.client.compute.ClientCompute;
 import org.apache.ignite.internal.client.io.ClientConnectionMultiplexer;
 import org.apache.ignite.internal.client.proto.ClientOp;
@@ -138,12 +137,6 @@ public class TcpIgniteClient implements IgniteClient {
     @Override
     public IgniteCompute compute() {
         return compute;
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public IgniteDeployment deployment() {
-        throw new UnsupportedOperationException("Deployment management not implemented yet.");
     }
 
     /** {@inheritDoc} */
