+++ modules/client/src/test/java/org/apache/ignite/client/fakes/FakeIgnite.java
@@ -22,7 +22,6 @@ import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 import org.apache.ignite.Ignite;
 import org.apache.ignite.compute.IgniteCompute;
-import org.apache.ignite.deployment.IgniteDeployment;
 import org.apache.ignite.internal.hlc.HybridTimestamp;
 import org.apache.ignite.internal.replicator.ReplicationGroupId;
 import org.apache.ignite.internal.sql.engine.QueryProcessor;
@@ -165,11 +164,6 @@ public class FakeIgnite implements Ignite {
     /** {@inheritDoc} */
     @Override
     public IgniteCompute compute() {
-        throw new UnsupportedOperationException("Not implemented yet");
-    }
-
-    @Override
-    public IgniteDeployment deployment() {
         throw new UnsupportedOperationException("Not implemented yet");
     }
