+++ modules/code-deployment/src/main/java/org/apache/ignite/internal/deployunit/DeployMetastoreService.java
@@ -30,8 +30,8 @@ import static org.apache.ignite.internal.metastorage.dsl.Operations.remove;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Flow.Publisher;
 import java.util.function.Consumer;
-import org.apache.ignite.deployment.version.Version;
 import org.apache.ignite.internal.deployunit.key.UnitMetaSerializer;
+import org.apache.ignite.internal.deployunit.version.Version;
 import org.apache.ignite.internal.logger.IgniteLogger;
 import org.apache.ignite.internal.logger.Loggers;
 import org.apache.ignite.internal.metastorage.Entry;
