+++ modules/code-deployment/src/main/java/org/apache/ignite/internal/deployunit/IgniteDeployment.java
@@ -15,16 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.deployment;
+package org.apache.ignite.internal.deployunit;
 
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import org.apache.ignite.deployment.version.Version;
+import org.apache.ignite.internal.deployunit.version.Version;
+import org.apache.ignite.internal.manager.IgniteComponent;
 
 /**
  * Provides access to the Deployment Unit functionality.
  */
-public interface IgniteDeployment {
+public interface IgniteDeployment extends IgniteComponent {
     /**
      * Deploy provided unit to current node.
      * After deploy finished, this deployment unit will be place to CMG group asynchronously.
