+++ modules/code-deployment/src/test/java/org/apache/ignite/deployment/UnitMetaSerializerTest.java
@@ -17,16 +17,16 @@
 
 package org.apache.ignite.deployment;
 
-import static org.apache.ignite.deployment.DeploymentStatus.UPLOADING;
+import static org.apache.ignite.internal.deployunit.DeploymentStatus.UPLOADING;
 import static org.apache.ignite.internal.deployunit.key.UnitMetaSerializer.deserialize;
 import static org.apache.ignite.internal.deployunit.key.UnitMetaSerializer.serialize;
 import static org.hamcrest.Matchers.is;
 
 import java.util.Arrays;
 import java.util.Collections;
-import org.apache.ignite.deployment.version.Version;
 import org.apache.ignite.internal.deployunit.UnitMeta;
 import org.apache.ignite.internal.deployunit.key.UnitMetaSerializer;
+import org.apache.ignite.internal.deployunit.version.Version;
 import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.Test;
