+++ modules/rest-api/openapi/openapi.yaml
@@ -327,7 +327,7 @@ paths:
       parameters: []
       responses:
         "200":
-          description: All statutes returned successfully.
+          description: All statuses returned successfully.
           content:
             application/json:
               schema:
@@ -402,7 +402,7 @@ paths:
           type: string
       responses:
         "200":
-          description: All statutes returned successfully.
+          description: All statuses returned successfully.
           content:
             application/json:
               schema:
