+++ modules/rest/src/main/java/org/apache/ignite/internal/rest/deployment/DeploymentManagementController.java
@@ -22,11 +22,16 @@ import io.micronaut.http.multipart.CompletedFileUpload;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
-import org.apache.ignite.deployment.DeploymentUnit;
-import org.apache.ignite.deployment.IgniteDeployment;
-import org.apache.ignite.deployment.version.Version;
+import org.apache.ignite.internal.deployunit.DeploymentUnit;
+import org.apache.ignite.internal.deployunit.IgniteDeployment;
+import org.apache.ignite.internal.deployunit.UnitStatus;
+import org.apache.ignite.internal.deployunit.version.Version;
 import org.apache.ignite.internal.rest.api.deployment.DeploymentCodeApi;
 import org.apache.ignite.internal.rest.api.deployment.UnitStatusDto;
 
@@ -66,7 +71,7 @@ public class DeploymentManagementController implements DeploymentCodeApi {
 
     @Override
     public CompletableFuture<Collection<UnitStatusDto>> units() {
-        return deployment.unitsAsync().thenApply(statuses -> statuses.stream().map(UnitStatusDto::fromUnitStatus)
+        return deployment.unitsAsync().thenApply(statuses -> statuses.stream().map(DeploymentManagementController::fromUnitStatus)
                 .collect(Collectors.toList()));
     }
 
@@ -78,13 +83,13 @@ public class DeploymentManagementController implements DeploymentCodeApi {
 
     @Override
     public CompletableFuture<UnitStatusDto> status(String unitId) {
-        return deployment.statusAsync(unitId).thenApply(UnitStatusDto::fromUnitStatus);
+        return deployment.statusAsync(unitId).thenApply(DeploymentManagementController::fromUnitStatus);
     }
 
     @Override
     public CompletableFuture<Collection<UnitStatusDto>> findByConsistentId(String consistentId) {
         return deployment.findUnitByConsistentIdAsync(consistentId)
-                .thenApply(units -> units.stream().map(UnitStatusDto::fromUnitStatus)
+                .thenApply(units -> units.stream().map(DeploymentManagementController::fromUnitStatus)
                         .collect(Collectors.toList()));
     }
 
@@ -102,5 +107,20 @@ public class DeploymentManagementController implements DeploymentCodeApi {
                 return is;
             }
         };
+    }
+
+    /**
+     * Mapper method.
+     *
+     * @param status Unit status.
+     * @return Unit status DTO.
+     */
+    public static UnitStatusDto fromUnitStatus(UnitStatus status) {
+        Map<String, List<String>> versionToConsistentIds = new HashMap<>();
+        Set<Version> versions = status.versions();
+        for (Version version : versions) {
+            versionToConsistentIds.put(version.render(), status.consistentIds(version));
+        }
+        return new UnitStatusDto(status.id(), versionToConsistentIds);
     }
 }
